move command helper files into directory
